SpecialGadgets: List right names instead of descriptions
The lengthy and localised descriptions make it hard to scan and
identify the actual rights themselves. They are meant as descriptions
to explain what a right is, not to identify the right itself.

It should probably be more similar to the list of checkboxes
on Special:UserRights. For now, kept the descriptions as tooltips,
but could also remove them entirely.

Change-Id: I29f68ffd5a8d002b65384f3762fb9e5631c13def