SpecialGadgets: Omit blank space due to <br> between <ul> and <p>
Due to some of the portions of the gadget list items being optional
it's not consistent whether or not they should be separated by
a line break.

When two pieces of text follow each other, there should be a line
break in-between.

When a piece of text is output afte a list, however, there is no
need for a line break as anything after a list would already
become its own line. Printing one anyway causes a blank space
to show up, which visually causes it to appear related to the
next gadget.

Fix by setting a boolean after each portion to indicate whether
or not it should be followed by a line break.

Also improve the code a bit with comments and use Html instead
of Xml.

Bug: T204616
Change-Id: If65e170746898999dd8f5351004eeaf49b340ab2