Use classic hook for CodeEditorGetPageLanguage
This hook needs the old call structur as there is not a interface for
it.

When CodeEditor extension is changed to provide that interface, this
needs to be moved to another class, because the extension is optional
and the interface not loaded on all installs.

This change is a follow-up to b504a6f4964bac74ec0bed817129c76700551c6b
and reverts the change for hook CodeEditorGetPageLanguage because
onCodeEditorGetPageLanguage is still a static function.

Bug: T271014
Change-Id: I0ac6f353be099c4c9b4cb36b65cb463b7b7b89ea