Avoid loading gadgets if the storage backend is disabled in tests
This method is possibly the main cause of DB access in non-database
tests, because it's called from the UserGetDefaultOptions hook handler,
which is triggered by most tests using UserOptionsLookup in any way.

Trying to obtain a DB connection will throw an exception in I96ecf9ff,
so just skip the lookup altogether.

Bug: T155147
Change-Id: I2ac659a472655b2a95e64dcbd3f709e465839b33