Streamline all kinds of different count() usages
None of these places does need the actual count. All they care about
is if an array is empty or not. Instead of converting an array to an
int and then to a bool we can go straight from array to bool.

Personally I believe this makes the code more consistent and more
readable.

Change-Id: Ic00b0740412c49a268b418d249aa4d3198e997f7