Rename category to section
In gadget preferences, the section under which a gadget is appears
is currently referred to as "category", which also appears in the JSON
page when using the JSON repo.

The introduction of categories-based gadgets in e2cb01dd causes
confusion as JSON pages now have both "categories" and "category"
which are completely unrelated.

This commit replaces occurrences of the term to "section". No backward
compatibility is provided as the JSON repo is not in use anywhere in
production.

Change-Id: I4309a5ee97fef7a9f16e2abb7579a8faf26549a3