Add gadgets to ApiParse result
Add gadgets via the ApiParseMakeOutputPage hook in addition to
BeforePageDisplay hook.

This adds default gadgets in the output for all pages for logged-out
invocations, resolving T161278.

As categories are added to the OutputPage instance before the
ApiParseMakeOutputPage hook is invoked, this also makes Live Preview
load category-triggered gadgets, resolving T376875.

Bug: T161278
Bug: T376875
Change-Id: Ie43b679dce6d48c6fdb5125bb125fa7ccf478c30